feat: add back to top scroll button
