+++ src/components/BackToTop.tsx
@@ -1,0 +1,51 @@
+import { useEffect, useState } from 'react';
+
+const SCROLL_THRESHOLD = 300;
+
+export default function BackToTop() {
+  const [isClient, setIsClient] = useState(false);
+  const [isVisible, setIsVisible] = useState(false);
+
+  useEffect(() => {
+    setIsClient(true);
+  }, []);
+
+  useEffect(() => {
+    if (!isClient || typeof window === 'undefined') {
+      return;
+    }
+
+    const handleScroll = () => {
+      setIsVisible(window.scrollY > SCROLL_THRESHOLD);
+    };
+
+    handleScroll();
+    window.addEventListener('scroll', handleScroll, { passive: true });
+
+    return () => {
+      window.removeEventListener('scroll', handleScroll);
+    };
+  }, [isClient]);
+
+  const handleClick = () => {
+    if (!isClient || typeof window === 'undefined') {
+      return;
+    }
+    window.scrollTo({ top: 0, behavior: 'smooth' });
+  };
+
+  if (!isClient || !isVisible) {
+    return null;
+  }
+
+  return (
+    <button
+      type="button"
+      onClick={handleClick}
+      aria-label="Back to top"
+      className="fixed bottom-6 right-6 z-50 h-12 w-12 rounded-full bg-black text-white shadow-lg transition hover:-translate-y-0.5 hover:bg-neutral-800 focus:outline-none focus-visible:ring-2 focus-visible:ring-black focus-visible:ring-offset-2"
+    >
+      <span aria-hidden="true" className="text-lg leading-none">↑</span>
+    </button>
+  );
+}
